Refactor NewsDownloader async download into httpasync module, also apply to Wikipedia
It'd have been simpler to just use Turbo really, but I don't want to risk those potential crashes on Android…
