+++ frontend/ui/wikipedia.lua
@@ -2,6 +2,7 @@ local JSON = require("json")
 local RenderImage = require("ui/renderimage")
 local Screen = require("device").screen
 local ffiutil = require("ffi/util")
+local httpasync = require("httpasync")
 local lfs = require("libs/libkoreader-lfs")
 local logger = require("logger")
 local time = require("ui/time")
@@ -1588,51 +1589,73 @@ abbr.abbr {
     -- ----------------------------------------------------------------
     -- OEBPS/images/*
     if include_images then
-        local nb_images = #images
         local before_images_time = time.now()
         local time_prev = before_images_time
+        local failed_images = {}
+
+        local tasks = {}
         for inum, img in ipairs(images) do
-            -- Process can be interrupted every second between image downloads
-            -- by tapping while the InfoMessage is displayed
-            -- We use the fast_refresh option from image #2 for a quicker download
-            local go_on
-            if time.to_ms(time.since(time_prev)) > 1000 then
-                time_prev = time.now()
-                go_on = UI:info(T(_("Retrieving image %1 / %2 …"), inum, nb_images), inum >= 2)
-                if not go_on then
-                    cancelled = true
-                    break
+            table.insert(tasks, {inum = inum, img = img})
+        end
+
+        local cancelled_download = httpasync.fetch_many(tasks, {
+            get_url = function(task)
+                local src = task.img.src
+                if use_img_2x and task.img.src2x then
+                    src = task.img.src2x
                 end
-            else
-                UI:info(T(_("Retrieving image %1 / %2 …"), inum, nb_images), inum >= 2, true)
-            end
-            local src = img.src
-            if use_img_2x and img.src2x then
-                src = img.src2x
-            end
-            logger.dbg("Getting img ", src)
-            local success, content = getUrlContent(src, nil, nil, true) -- reuse_cookie
-            -- success, content = getUrlContent(src..".unexistant") -- to simulate failure
-            if success then
-                logger.dbg("success, size:", #content)
-            else
-                logger.info("failed fetching:", src)
-            end
-            if success then
+                return src
+            end,
+            fetch = function(src)
+                logger.dbg("Getting img async:", src)
+                local success, _, content = httpasync.fetch_url(src, 0, {
+                    cookie = cached_cookie,
+                    referer = "https://en.wikipedia.org/",
+                })
+                if not success then
+                    -- Fallback to the synchronous getter for edge cases.
+                    success, content = getUrlContent(src, nil, nil, true) -- reuse_cookie
+                end
+                return success, content
+            end,
+            on_success = function(task, content)
                 -- Images do not need to be compressed, so spare some cpu cycles
-                if img.mimetype ~= "image/svg+xml" then -- except for SVG images (which are XML text)
+                if task.img.mimetype ~= "image/svg+xml" then -- except for SVG images (which are XML text)
                     epub:setZipCompression("store")
                 end
-                epub:addFileFromMemory("OEBPS/"..img.imgpath, content, mtime)
+                epub:addFileFromMemory("OEBPS/"..task.img.imgpath, content, mtime)
                 epub:setZipCompression("deflate")
-            else
-                go_on = UI:confirm(T(_("Downloading image %1 failed. Continue anyway?"), inum), _("Stop"), _("Continue"))
-                if not go_on then
-                    cancelled = true
-                    break
+            end,
+            on_failure = function(task, err)
+                logger.info("failed fetching:", task.img.src, err)
+                table.insert(failed_images, task.inum)
+            end,
+            on_progress = function(completed, total)
+                -- Process can be interrupted every second by tapping while the
+                -- InfoMessage is displayed.
+                if time.to_ms(time.since(time_prev)) > 1000 then
+                    time_prev = time.now()
+                    local go_on = UI:info(T(_("Retrieving images… %1 / %2 completed"), completed, total), completed >= 1)
+                    if not go_on then
+                        return false -- cancel
+                    end
                 end
+                return true
+            end,
+        })
+
+        if cancelled_download then
+            cancelled = true
+        end
+
+        -- Report any failures once, rather than interrupting on each one.
+        if not cancelled and #failed_images > 0 then
+            local go_on = UI:confirm(T(_("%1 images failed to download. Continue creating the EPUB?"), #failed_images), _("Stop"), _("Continue"))
+            if not go_on then
+                cancelled = true
             end
         end
+
         logger.dbg("Image download time for:", page_htmltitle, time.to_ms(time.since(before_images_time)), "ms")
     end
